requirements.txt
customtkinter>=5.2.0
psutil>=5.9.0
scapy>=2.5.0
python-nmap>=0.7.1
netifaces>=0.11.0
requests>=2.31.0
mac-vendor-lookup>=2.1.0
pyudev>=0.24.0
